Feat : Added new enemy type : crawler
Abilities : travels at 1.5 times speed of flare, has same health as
mono, and is invisible, unless hit by any status effect. Once hit, it becomes visible for the
period that the status effect remains. The status effect applies it's
own debuffs

Turrets cannot aim at it unless it is visible, so it must be used with
effect applying turrets.